requirements.txt
python-dotenv==1.0.1
langchain-openai==0.3.9